% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar functions.R
\name{locstep}
\alias{locstep}
\title{Local linear regression with stepwise forward and maximum number of features}
\usage{
locstep(arg, x, y, h, dmax, varimp)
}
\description{
Local linear regression with stepwise forward and maximum number of features
}
