% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{ploteffect}
\alias{ploteffect}
\title{Plot VarImp effect}
\usage{
ploteffect(instancen, x, y, xcoef)
}
\arguments{
\item{instancen}{Instance index.}

\item{x}{Matrix \eqn{n \times d} with covariates for each sample unit.}

\item{y}{\eqn{n}-dimensional vector with predictions from the original model.}

\item{xcoef}{Matrix \eqn{n \times d} with local slopes for each sample unit.}
}
\value{
Graph with effect of one instance.
}
\description{
This function plot VarImp effect of one instance.
}
\examples{
library(interpretability)

data("artificial_data")

x_coef <- varimp(x = artificial_data$x_test,
                 y = artificial_data$y_pred,
                 h = 0.1,
                 dmax = ncol(artificial_data$x_test))

ploteffect(instancen = 1,
          x = artificial_data$x_test,
          y = artificial_data$y_pred,
          xcoef = x_coef)
}
