% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{ploteffectclus}
\alias{ploteffectclus}
\title{Plot SupClus effect}
\usage{
ploteffectclus(instancen, x, y, clustercoef, cluster)
}
\arguments{
\item{instancen}{Instance index.}

\item{x}{Matrix \eqn{n \times d} with covariates for each sample unit.}

\item{y}{\eqn{n}-dimensional vector with predictions from the original model.}

\item{clustercoef}{Matrix \eqn{k x d} with local slopes for each cluster.}

\item{cluster}{\eqn{n}-dimensional vector with cluster of each instance.}
}
\value{
Graph with effect of one instance.
}
\description{
This function plot SupClus effect of one instance.
}
\examples{
library(interpretability)

data("artificial_data")

supcluscoef <- supclus(k = 2,
                       x = artificial_data$x_test,
                       y = artificial_data$y_pred,
                       h = 0.1,
                       dmax = ncol(artificial_data$x_test))
                       
ploteffectclus(instancen = 1,
              x = artificial_data$x_test,
              y = artificial_data$y_pred, 
              clustercoef = supcluscoef$coef,  
              cluster = supcluscoef$cluster)
}
