% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main functions.R
\name{plotslopes}
\alias{plotslopes}
\title{Plot slopes}
\usage{
plotslopes(sampsize, varnumber, x, y, xcoef)
}
\arguments{
\item{sampsize}{Sample size for plotting slopes.}

\item{varnumber}{Covariate index.}

\item{x}{Matrix \eqn{n \times d} with covariates for each sample unit.}

\item{y}{\eqn{n}-dimensional vector with predictions from the original model.}

\item{xcoef}{Matrix \eqn{n x d} with local slopes for each sample unit.}
}
\value{
Graph with local slopes of some points for one feature.
}
\description{
This function plot local slopes of some points for one feature.
}
\examples{
library(interpretability)

data("artificial_data")

x_coef <- varimp(x = artificial_data$x_test,
                 y = artificial_data$y_pred,
                 h = 0.1,
                 dmax = ncol(artificial_data$x_test))

plotslopes(sampsize = 50,
           varnumber = 1,
           x = artificial_data$x_test,
           y = artificial_data$y_pred,
           xcoef = x_coef)
           
}
