% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar functions.R
\name{regcluster}
\alias{regcluster}
\title{Create cluster with coefficients and fit regressions for each cluster}
\usage{
regcluster(k, xcoef, x, y, dmax = ncol(x))
}
\description{
Create cluster with coefficients and fit regressions for each cluster
}
